+++ examples/cra/src/App.tsx
@@ -1,12 +1,12 @@
 import './App.css'
-import { SignOrderPage } from './components/signOrder'
-import { SignOrderCancellationPage } from './components/signOrderCancellation'
-import { GetTradesPage } from './components/getTrades'
-import { GetOrdersPage } from './components/getOrders'
-import { GetQuotePage } from './components/getQuote'
-import { SignAndSendOrderPage } from './components/sendOrder'
-import { SendOrderCancellationPage } from './components/sendOrderCancellation'
-import { createContext, FC, useEffect, useState } from 'react'
+import { SignOrderPage } from './pages/signOrder'
+import { SignOrderCancellationPage } from './pages/signOrderCancellation'
+import { GetTradesPage } from './pages/getTrades'
+import { GetOrdersPage } from './pages/getOrders'
+import { GetQuotePage } from './pages/getQuote'
+import { SignAndSendOrderPage } from './pages/sendOrder'
+import { SendOrderCancellationPage } from './pages/sendOrderCancellation'
+import { FC, useEffect, useState } from 'react'
 import { useWeb3Info } from './hooks/useWeb3Info'
 import { ChainIdContext } from './context'
