+++ examples/cra/src/pages/sendOrder/index.tsx
@@ -2,8 +2,8 @@ import { FormEvent, useCallback, useEffect, useState } from 'react'
 import '../../pageStyles.css'
 import { OrderSigningUtils, UID, OrderKind, UnsignedOrder, OrderBookApi, SigningScheme } from '@cowprotocol/cow-sdk'
 import { useWeb3Info } from '../../hooks/useWeb3Info'
-import { JsonContent } from '../jsonContent'
-import { ResultContent } from '../resultContent'
+import { JsonContent } from '../../components/jsonContent'
+import { ResultContent } from '../../components/resultContent'
 import { useCurrentChainId } from '../../hooks/useCurrentChainId'
 
 const orderBookApi = new OrderBookApi()
