+++ examples/cra/src/pages/signOrderCancellation/index.tsx
@@ -3,8 +3,8 @@ import '../../pageStyles.css'
 import { OrderCancellations, OrderSigningUtils, SigningResult } from '@cowprotocol/cow-sdk'
 import { useWeb3Info } from '../../hooks/useWeb3Info'
 import { parseFormData } from '../../utils'
-import { JsonContent } from '../jsonContent'
-import { ResultContent } from '../resultContent'
+import { JsonContent } from '../../components/jsonContent'
+import { ResultContent } from '../../components/resultContent'
 import { useCurrentChainId } from '../../hooks/useCurrentChainId'
 import { signOrderCancellation } from '@cowprotocol/contracts'
